requirements.txt
# Rustic Garden — Azure Function App Dependencies
# Deploy with: pip install -r requirements.txt

azure-functions==1.21.3
